requirements.txt
Django>=4
django-extensions
djangorestframework
django-htmx
